XRDDEV-962 refactor, extract ServiceClientServiceTest, make AccessRightServiceTest not depend on others
